rename AudioFile/{AudioFileExtendedDouble => extended_double}.hpp
