+++ include/AudioFile/OAudioFile.hpp
@@ -2,7 +2,7 @@
 #define _HISSTOOLS_OAUDIOFILE_
 
 #include "AudioFileUtilities.hpp"
-#include "AudioFileExtendedDouble.hpp"
+#include "extended_double.hpp"
 #include "BaseAudioFile.hpp"
 
 #include <algorithm>
