+++ include/AudioFile/extended_double.hpp
